+++ src/RcppExports.cpp
@@ -31,6 +31,17 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
+// alloc_non_r_bytes
+Rcpp::XPtr<std::vector<unsigned char>> alloc_non_r_bytes(long n_bytes);
+RcppExport SEXP _clustermq_alloc_non_r_bytes(SEXP n_bytesSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< long >::type n_bytes(n_bytesSEXP);
+    rcpp_result_gen = Rcpp::wrap(alloc_non_r_bytes(n_bytes));
+    return rcpp_result_gen;
+END_RCPP
+}
 
 RcppExport SEXP _rcpp_module_boot_cmq_master();
 RcppExport SEXP _rcpp_module_boot_cmq_proxy();
@@ -39,6 +50,7 @@ RcppExport SEXP _rcpp_module_boot_cmq_worker();
 static const R_CallMethodDef CallEntries[] = {
     {"_clustermq_has_connectivity", (DL_FUNC) &_clustermq_has_connectivity, 1},
     {"_clustermq_libzmq_has_draft", (DL_FUNC) &_clustermq_libzmq_has_draft, 0},
+    {"_clustermq_alloc_non_r_bytes", (DL_FUNC) &_clustermq_alloc_non_r_bytes, 1},
     {"_rcpp_module_boot_cmq_master", (DL_FUNC) &_rcpp_module_boot_cmq_master, 0},
     {"_rcpp_module_boot_cmq_proxy", (DL_FUNC) &_rcpp_module_boot_cmq_proxy, 0},
     {"_rcpp_module_boot_cmq_worker", (DL_FUNC) &_rcpp_module_boot_cmq_worker, 0},