+++ src/memory.h
@@ -1,0 +1,8 @@
+#ifndef _MEMORY_H_
+#define _MEMORY_H_
+
+#include <Rcpp.h>
+
+Rcpp::NumericVector mem_stats();
+
+#endif // _MEMORY_H_